Move delete_workspace_publication(s) to base class
